+++ orix/plot/__init__.py
@@ -28,14 +28,13 @@ Example of usage::
     >>> ax.scatter(vector.Vector3d([[0, 0, 1], [1, 0, 1]]))
 """
 
-from orix.plot.crystal_map_plot import convert_unit, CrystalMapPlot
+from orix.plot.crystal_map_plot import CrystalMapPlot
 from orix.plot.rotation_plot import AxAnglePlot, RodriguesPlot, RotationPlot
 from orix.plot.stereographic_plot import StereographicPlot
 
 
 # Lists what will be imported when calling "from orix.plot import *"
 __all__ = [
-    "convert_unit",
     "CrystalMapPlot",
     "AxAnglePlot",
     "RodriguesPlot",